+++ packages/zowe-explorer/src/uss/actions.ts
@@ -64,7 +64,6 @@ export async function createUSSNode(node: IZoweUSSTreeNode, ussFileProvider: IZo
             filePath = `${filePath}/${name}`;
             await ZoweExplorerApiRegister.getUssApi(node.getProfile()).create(filePath, nodeType);
             if (isTopLevel) {
-                await Profiles.getInstance().refresh(ZoweExplorerApiRegister.getInstance());
                 refreshAll(ussFileProvider);
             } else {
                 ussFileProvider.refreshElement(node);